chore(conf) worker_rlimit_nofile & worker_connections "auto" to 65536
### Summary

This PR descriprion already summarizes that default `auto` value for
`nginx_main_worker_rlimit_nofile` and `nginx_events_worker_connections`
should have been set to `65536`. But it seems like it was still capped
at `16384`.

Not sure what happened there. Was it a last-minute change or did we
just made mistake originally. This PR now changes that default value
to `65536`, and changes a test to check that it is explicitly that.

I also decided to document this in `kong.conf.default` which is used
for auto-generating the docs for configuration parameters.